+++ src/main/java/de/dm/intellij/liferay/language/sass/LiferayFrontendTokenDefinitionVariableReference.java
@@ -1,9 +1,11 @@
 package de.dm.intellij.liferay.language.sass;
 
+import com.intellij.codeInsight.lookup.LookupElementBuilder;
 import com.intellij.codeInspection.util.InspectionMessage;
 import com.intellij.json.psi.JsonFile;
 import com.intellij.json.psi.JsonProperty;
 import com.intellij.json.psi.JsonValue;
+import com.intellij.navigation.ItemPresentation;
 import com.intellij.openapi.module.Module;
 import com.intellij.openapi.module.ModuleUtil;
 import com.intellij.openapi.vfs.VirtualFile;
@@ -13,10 +15,13 @@ import com.intellij.psi.PsiManager;
 import com.intellij.psi.PsiReferenceBase;
 import com.intellij.psi.css.CssBundle;
 import com.intellij.psi.css.reference.CssReference;
+import com.intellij.util.IncorrectOperationException;
 import de.dm.intellij.liferay.util.LiferayFileUtil;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 
 public class LiferayFrontendTokenDefinitionVariableReference extends PsiReferenceBase<PsiElement> implements CssReference {
@@ -34,6 +39,60 @@ public class LiferayFrontendTokenDefinitionVariableReference extends PsiReferenc
 
     @Override
     public @Nullable PsiElement resolve() {
+        Map<String, JsonProperty> cssVariableMap = getCssVariableMap();
+
+        if (cssVariableMap != null) {
+            String text = getElement().getText();
+
+            if (text.length() > 2) {
+                if (text.startsWith("--")) {
+                    String variableName = text.substring(2);
+
+                    JsonProperty jsonProperty = cssVariableMap.get(variableName);
+
+                    if (jsonProperty != null) {
+                        return jsonProperty.getValue();
+                    }
+                }
+            }
+        }
+
+        return null;
+    }
+
+    @Override
+    public Object @NotNull [] getVariants() {
+        List<Object> result = new ArrayList<Object>();
+
+        Map<String, JsonProperty> cssVariableMap = getCssVariableMap();
+
+        if (cssVariableMap != null) {
+            for (String cssVariable : cssVariableMap.keySet()) {
+                JsonProperty jsonProperty = cssVariableMap.get(cssVariable);
+
+                ItemPresentation presentation = jsonProperty.getPresentation();
+
+                if (presentation != null && jsonProperty.getValue() != null) {
+                    result.add(
+                        LookupElementBuilder.createWithSmartPointer(
+                                "--" + cssVariable,
+                                jsonProperty.getValue()
+                        )
+                            .withIcon(presentation.getIcon(false))
+                            .withTypeText(presentation.getLocationString(), true)
+                    );
+                }
+            }
+        }
+
+        return result.toArray(new Object[0]);
+    }
+
+    public PsiElement handleElementRename(@NotNull String newElementName) throws IncorrectOperationException {
+        return super.handleElementRename(newElementName.startsWith("--") ? newElementName : "--" + newElementName);
+    }
+
+    private Map<String, JsonProperty> getCssVariableMap() {
         final Module module = ModuleUtil.findModuleForPsiElement(getElement());
 
         if (module != null) {
@@ -54,17 +113,7 @@ public class LiferayFrontendTokenDefinitionVariableReference extends PsiReferenc
 
                         topLevelValue.accept(liferayFrontendTokenDefinitionJsonElementVisitor);
 
-                        Map<String, JsonProperty> cssVariableMap = liferayFrontendTokenDefinitionJsonElementVisitor.getCssVariableMap();
-
-                        String text = getElement().getText();
-
-                        if (text.length() > 2) {
-                            if (text.startsWith("--")) {
-                                String variableName = text.substring(2);
-
-                                return cssVariableMap.get(variableName);
-                            }
-                        }
+                        return liferayFrontendTokenDefinitionJsonElementVisitor.getCssVariableMap();
                     }
                 }
             }