+++ changelog.md
@@ -7,8 +7,10 @@ All notable changes to this project will be documented in this file.
 - Feature: Added support for Liferay 7.4
 - Feature: Code Completion and documentation for portal.properties files
 - Feature: JSON Schema for frontend-token-definition.json
+- Feature: Language.properties references for label attributes in frontend-token-definition.json  
 - Feature: Clay SCSS Variables documentation
-
+- Bugfix: Fixed detecting the bundles directory from the bundle-support maven plugin
+  
 ## 0.0.11
 - Feature: Implicit usage of some property files and keys ("javax.portlet.title." or "liferay.workspace."), so that they are no longer shown as "unused property"
 - Feature: Inspection of @Reference usage outside of a @Component class